+++ run_coverage.sh
@@ -1,0 +1,4 @@
+#!/bin/sh -
+python -m coverage run --source ojax -m unittest discover -s tests/ -v
+python -m coverage report
+python -m coverage html