+++ tests/test_ojax.py
@@ -1,5 +1,8 @@
 import unittest
 import typing
+import abc
+import warnings
+import io
 import dataclasses
 import jax
 import ojax
@@ -11,7 +14,34 @@ class TestOTree(unittest.TestCase):
             class MissingAnno(ojax.OTree):
                 my_field = 42
 
-            ma = MissingAnno()
+            _ = MissingAnno()
+
+    def test_warn_no_annotation_abc(self):
+        with warnings.catch_warnings():
+            warnings.simplefilter("error")
+
+            class AbstractTree(ojax.OTree, metaclass=abc.ABCMeta):
+                @abc.abstractmethod
+                def my_method(self):
+                    raise NotImplementedError
+
+            class MyTree(AbstractTree):
+                def my_method(self):
+                    return 42
+
+            mt = MyTree()
+            self.assertEqual(mt.my_method(), 42)
+
+    def test_warn_no_annotation_property(self):
+        with warnings.catch_warnings():
+            warnings.simplefilter("error")
+
+            class WithProperty(ojax.OTree):
+                @property
+                def my_property(self):
+                    return 42
+
+            _ = WithProperty()
 
     def test_field_type(self):
         class MyOTree(ojax.OTree):
@@ -20,15 +50,23 @@ class TestOTree(unittest.TestCase):
             c: ojax.OTree = ojax.ignore()
             d: typing.ClassVar[jax.Array]
             e: dataclasses.InitVar[str]
+            f: io.BytesIO = ojax.alien()
 
         ftmap = {
             'a': ojax.Aux,
             'b': ojax.Child,
             'c': ojax.Ignore,
+            'f': ojax.Alien,
         }
-        self.assertEqual(len(tuple(dataclasses.fields(MyOTree))), 3)
+        self.assertEqual(len(tuple(dataclasses.fields(MyOTree))), 4)
         for f in dataclasses.fields(MyOTree):
             self.assertTrue(isinstance(f, ftmap[f.name]))
+        for f in ojax.fields(MyOTree):
+            self.assertTrue(isinstance(f, ftmap[f.name]))
+        for f in ojax.fields(MyOTree, ojax.Ignore, infer=False):
+            self.assertTrue(isinstance(f, ftmap[f.name]))
+        with self.assertRaises(ValueError):
+            _ = ojax.fields(MyOTree, dataclasses.Field)
 
     def test_aux_infer(self):
         class MyOTree(ojax.OTree):
@@ -61,14 +99,51 @@ class TestOTree(unittest.TestCase):
             c: ojax.OTree = ojax.ignore()
             e: dataclasses.InitVar[str]
             d: typing.ClassVar[typing.Any] = None
+            f: io.BytesIO = ojax.alien(default_factory=io.BytesIO)
 
             def __post_init__(self, e) -> None:
                 assert e == 'lol'
+                super().__post_init__()
 
         try:
-            my_otree = MyOTree(a=0, b=4.2, c=None, e='lol')
+            _ = MyOTree(a=0, b=4.2, c=None, e='lol')
         except:
             self.fail("OTree initialization failed unexpectedly.")
+
+    def test_flatten_unflatten(self):
+        class MyOTree(ojax.OTree):
+            a: jax.Array = ojax.aux()
+            b: float = ojax.child()
+            c: typing.Any = ojax.ignore()
+            e: dataclasses.InitVar[str]
+            d: typing.ClassVar[typing.Any] = None
+
+            def __post_init__(self, e) -> None:
+                assert e == 'lol'
+                super().__post_init__()
+
+        my_otree = MyOTree(a=0, b=4.2, c=None, e='lol')
+        try:
+            vals, treedef = jax.tree.flatten(my_otree)
+        except:
+            self.fail("OTree flatten failed unexpectedly.")
+        try:
+            new_otree = jax.tree.unflatten(treedef, vals)
+        except:
+            self.fail("OTree unflatten failed unexpectedly.")
+        for n in ('a', 'b'):
+            self.assertEqual(getattr(my_otree, n), getattr(new_otree, n))
+
+        class AlienOTree(MyOTree):
+            f: io.BytesIO = ojax.alien(default_factory=io.BytesIO)
+
+            def add_b(self, val):
+                return val + self.b
+
+        alien_otree = AlienOTree(a=0, b=4.2, c=None, e='lol')
+        self.assertEqual(jax.jit(alien_otree.add_b)(10), 14.2)
+        with self.assertRaises(ojax.AlienException):
+            _ = jax.tree.flatten(alien_otree)
 
     def test_update(self):
         class MyOTree(ojax.OTree):
@@ -77,6 +152,7 @@ class TestOTree(unittest.TestCase):
             c: typing.Any = ojax.ignore()
             e: dataclasses.InitVar[str]
             d: typing.ClassVar[typing.Any] = None
+            f: io.BytesIO = ojax.alien(default_factory=io.BytesIO)
 
             def __post_init__(self, e) -> None:
                 assert e == 'lol'
@@ -93,6 +169,13 @@ class TestOTree(unittest.TestCase):
         except:
             self.fail("OTree update failed unexpectedly.")
         self.assertEqual(my_otree.b, 42.0)
+        try:
+            my_otree = my_otree.update(
+                f=None
+            )
+        except:
+            self.fail("OTree update failed unexpectedly.")
+        self.assertEqual(my_otree.f, None)
         # changing aux should fail
         with self.assertRaises(ValueError):
             my_otree.update(a=1)